move to `electron-main` folder
