+++ electron-main/getFileAndFolder.js
