+++ electron-main/logOperations.js
