+++ electron-main/videoFileOperations.js
